+++ src/apiflask/types.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 import typing as t
+from abc import abstractmethod
 from typing import Protocol
+from typing import runtime_checkable
 
 if t.TYPE_CHECKING:  # pragma: no cover
     from flask.wrappers import Response  # noqa: F401
@@ -11,6 +13,9 @@ if t.TYPE_CHECKING:  # pragma: no cover
     from .schemas import Schema  # noqa: F401
     from .security import HTTPBasicAuth  # noqa: F401
     from .security import HTTPTokenAuth  # noqa: F401
+    from .security import APIKeyHeaderAuth  # noqa: F401
+    from .security import APIKeyCookieAuth  # noqa: F401
+    from .security import APIKeyQueryAuth  # noqa: F401
     from .exceptions import HTTPError  # noqa: F401
     from .views import View  # noqa: F401
 
@@ -51,7 +56,9 @@ ErrorCallbackType = t.Callable[['HTTPError'], ResponseReturnValueType]
 DictSchemaType = t.Dict[str, t.Union['Field', type]]
 SchemaType = t.Union['Schema', t.Type['Schema'], DictSchemaType]
 OpenAPISchemaType = t.Union['Schema', t.Type['Schema'], dict]
-HTTPAuthType = t.Union['HTTPBasicAuth', 'HTTPTokenAuth']
+HTTPAuthType = t.Union[
+    'HTTPBasicAuth', 'HTTPTokenAuth', 'APIKeyHeaderAuth', 'APIKeyCookieAuth', 'APIKeyQueryAuth'
+]
 TagsType = t.Union[t.List[str], t.List[t.Dict[str, t.Any]]]
 ViewClassType = t.Type['View']
 ViewFuncOrClassType = t.Union[t.Callable, ViewClassType]
@@ -80,3 +87,14 @@ class PaginationType(Protocol):
 class ViewFuncType(Protocol):
     _spec: t.Any
     _method_spec: t.Any
+
+
+@runtime_checkable
+class SecurityScheme(Protocol):
+    name: str
+
+    @abstractmethod
+    def get_security_scheme(self) -> dict[str, t.Any]:
+        """Get security scheme."""
+
+        pass  # pragma: no cover