+++ CHANGES.md
@@ -6,9 +6,15 @@ Released: -
 - Drop support for Python 3.8 and PyPy 3.10.
 - Decouple from marshmallow and add an adapter system to support different serialization/deserialization libraries ([pr #690](pr_690)).
 - Add support for Pydantic models as data schemas ([issue #519][issue_519]).
+- Fix subclassed MethodView resources cannot be added as URL rules ([issue #618][issue_618]).
+- Add support for API key auth with `APIKeyHeaderAuth`, `APIKeyCookieAuth`, and `APIKeyQueryAuth`. Add support for runtime selection of authentication methods with `MultiAuth`. Deprecate the API key auth with HTTPTokenAuth ([issue #604][issue_604]).
+- Removed implicit security scheme naming rules ([pr #665](pr_665)).
 
 [issue_519]: https://github.com/apiflask/apiflask/issues/519
 [pr_690]: https://github.com/apiflask/apiflask/pull/690
+[issue_618]: https://github.com/apiflask/apiflask/issues/618
+[issue_604]: https://github.com/apiflask/apiflask/issues/604
+[pr_655]: https://github.com/apiflask/apiflask/pull/655
 
 
 ## Version 2.4.0